refactor(client): renames files
